% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESubgrp_Gcom.R
\name{TESubgrp_Gcom}
\alias{TESubgrp_Gcom}
\title{regression tree}
\description{
regression tree

regression tree
}
\section{Super class}{
\code{\link[IntpTrtRule:TEsubgrp]{IntpTrtRule::TEsubgrp}} -> \code{TEsubgrp_Gcom}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TESubgrp_Gcom$new()}}
\item \href{#method-clone}{\code{TESubgrp_Gcom$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="IntpTrtRule" data-topic="TEsubgrp" data-id="PlotCostSensTree">}\href{../../IntpTrtRule/html/TEsubgrp.html#method-PlotCostSensTree}{\code{IntpTrtRule::TEsubgrp$PlotCostSensTree()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="IntpTrtRule" data-topic="TEsubgrp" data-id="SankeyNetworkPlot">}\href{../../IntpTrtRule/html/TEsubgrp.html#method-SankeyNetworkPlot}{\code{IntpTrtRule::TEsubgrp$SankeyNetworkPlot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="IntpTrtRule" data-topic="TEsubgrp" data-id="fit">}\href{../../IntpTrtRule/html/TEsubgrp.html#method-fit}{\code{IntpTrtRule::TEsubgrp$fit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="IntpTrtRule" data-topic="TEsubgrp" data-id="print_leaf_nodes">}\href{../../IntpTrtRule/html/TEsubgrp.html#method-print_leaf_nodes}{\code{IntpTrtRule::TEsubgrp$print_leaf_nodes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TESubgrp_Gcom$new(x, trt, y1, y0, y, xvars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a dataframe of patient-level covariates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TESubgrp_Gcom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
